+++ sphinx/ext/apidoc/__init__.py
@@ -13,9 +13,27 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+import sphinx
 from sphinx.ext.apidoc._cli import main
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
 
-__all__: Sequence[str] = ('main',)
+    from sphinx.application import Sphinx
+    from sphinx.util.typing import ExtensionMetadata
+
+__all__: Sequence[str] = 'main', 'setup'
+
+
+def setup(app: Sphinx) -> ExtensionMetadata:
+    from sphinx.ext.apidoc._extension import run_apidoc
+
+    # Require autodoc
+    app.setup_extension('sphinx.ext.autodoc')
+    app.add_config_value('apidoc_defaults', {}, 'env', types=frozenset({dict}))
+    app.add_config_value('apidoc_modules', (), 'env', types=frozenset((list, tuple)))
+    app.connect('builder-inited', run_apidoc)
+    return {
+        'version': sphinx.__display_version__,
+        'parallel_read_safe': True,
+    }