+++ tests/test_extensions/test_ext_apidoc.py
@@ -13,6 +13,8 @@ from sphinx.ext.apidoc._cli import main as apidoc_main
 if TYPE_CHECKING:
     from pathlib import Path
 
+    from sphinx.testing.util import SphinxTestApp
+
 _apidoc = namedtuple('_apidoc', 'coderoot,outdir')  # NoQA: PYI024
 
 
@@ -771,3 +773,26 @@ def test_remove_old_files(tmp_path: Path):
     apidoc_main(['--remove-old', '-o', str(gen_dir), str(module_dir)])
     assert set(gen_dir.iterdir()) == {gen_dir / 'modules.rst', gen_dir / 'example.rst'}
     assert (gen_dir / 'example.rst').stat().st_mtime_ns == example_mtime
+
+
+@pytest.mark.sphinx(testroot='ext-apidoc')
+def test_sphinx_extension(app: SphinxTestApp):
+    """Test running apidoc as an extension."""
+    app.build()
+    assert app.warning.getvalue() == ''
+
+    assert set((app.srcdir / 'generated').iterdir()) == {
+        app.srcdir / 'generated' / 'modules.rst',
+        app.srcdir / 'generated' / 'my_package.rst',
+    }
+    assert 'show-inheritance' not in (
+        app.srcdir / 'generated' / 'my_package.rst'
+    ).read_text(encoding='utf8')
+    assert (app.outdir / 'generated' / 'my_package.html').is_file()
+
+    # test a re-build
+    app.build()
+    assert app.warning.getvalue() == ''
+
+    # TODO check nothing got re-built
+    # TODO test that old files are removed