+++ CHANGES.rst
@@ -106,6 +106,10 @@ Features added
 * #9169: Add the :confval:`intersphinx_resolve_self` option
   to resolve an intersphinx reference to the current project.
   Patch by Jakob Lykke Andersen and Adam Turner.
+* #13333: Add the :mod:`sphinx.ext.autodoc` extension,
+  to automate API documentation generation
+  from Python modules.
+  Patch by Chris Sewell and Adam Turner.
 
 Bugs fixed
 ----------